Move gridpaneConstraints()  to Node.kt
