+++ src/main/java/tornadofx/Builders.kt
@@ -2,9 +2,6 @@ package tornadofx
 
 import javafx.beans.property.Property
 import javafx.collections.ObservableList
-import javafx.geometry.HPos
-import javafx.geometry.Insets
-import javafx.geometry.VPos
 import javafx.scene.Node
 import javafx.scene.control.*
 import javafx.scene.image.ImageView
@@ -121,62 +118,6 @@ fun Pane.vbox(spacing: Double? = null, children: Iterable<Node>? = null, op: (VB
 
 fun Pane.stackpane(initialChildren: Iterable<Node>? = null, op: (StackPane.() -> Unit)? = null) = opcr(this, StackPane().apply { if (initialChildren != null) children.addAll(initialChildren) }, op)
 fun Pane.gridpane(op: (GridPane.() -> Unit)? = null) = opcr(this, GridPane(), op)
-
-
-fun <T: Node> T.gridpaneConstraints(op: (GridPaneConstraint.() -> Unit)): T {
-    val gpc = GridPaneConstraint()
-    gpc.op()
-    return gpc.applyToNode(this)
-}
-
-class GridPaneConstraint(
-        var columnIndex: Int? = null,
-        var rowIndex: Int? = null,
-        var hGrow: Priority? = null,
-        var vGrow: Priority? = null,
-        var margin: Insets? = null,
-        var fillHeight: Boolean? = null,
-        var fillWidth: Boolean? = null,
-        var hAlignment: HPos? = null,
-        var vAlignment: VPos? = null
-)
-{
-    var vhGrow: Priority? = null
-        set(value) {
-            vGrow = value
-            hGrow = value
-            field = value
-        }
-
-    var fillHeightWidth: Boolean? = null
-        set(value) {
-            fillHeight = value
-            fillWidth = value
-            field = value
-        }
-
-    fun columnRowIndex(columIndex: Int, rowIndex: Int) {
-        this.columnIndex = columnIndex
-        this.rowIndex = rowIndex
-    }
-    fun fillHeightWidth(fill: Boolean) {
-        fillHeight = fill
-        fillWidth = fill
-    }
-
-    fun <T: Node> applyToNode(node: T): T {
-        columnIndex?.let { GridPane.setColumnIndex(node,it) }
-        rowIndex?.let { GridPane.setRowIndex(node,it) }
-        hGrow?.let { GridPane.setHgrow(node,it) }
-        vGrow?.let { GridPane.setVgrow(node,it) }
-        margin?.let { GridPane.setMargin(node,it) }
-        fillHeight?.let { GridPane.setFillHeight(node,it) }
-        fillWidth?.let { GridPane.setFillWidth(node,it) }
-        hAlignment?.let { GridPane.setHalignment(node,it) }
-        vAlignment?.let { GridPane.setValignment(node,it) }
-        return node
-    }
-}
 
 /**
  * Add the given node to the pane, invoke the node operation and return the node
